Clean up --help, add stub for initial setup helper.
Big thanks to the sysadmin team for setting up URL shortlinks for the
docs URLs!

This takes another step towards the idea of being able to automatically
install known dependencies for common distributions, adding .xsession,
adjusting .bashrc, making a configuration file, and all the other
initial tasks that need to be done.