first-run: Make it possible to run system installer on first run.
Tested with OpenSuSE docker image (opensuse-leap:latest). With it, this
sequence of commands run in the container succeeds, with kdesrc-build
git dir bind-mounted to /opt/kdesrc-build in the container:

    export PATH=/opt/kdesrc-build:$PATH
    kdesrc-build --initial-setup
    kdesrc-build --metadata-only
    kdesrc-build extra-cmake-modules

I mention containers, but that's mostly to make it easier to test. There
is separate work by others to actually make kdesrc-build part of a fully
container-based workflow which I don't want/need to reinvent.