mojo: Output cleanup, fix ksb::Debug::error.
All ksb::Debug output methods when called in a child subprocess are
supposed to send the log to the parent process. This wasn't working for
ksb::Debug::error because it sent output to STDERR directly instead of
using print_clr. But in the subprocess both STDOUT and STDERR are
redirected away from the TTY.

In the meantime I've output the error log for a build failure as soon as
it's available instead of waiting for the end. To avoid duplicates I
don't print any of the per-module info at the end now. Maybe there's a
better way to do this?