doc: Add source reference documentation for ksb::Module.
I already have "how to understand the source" docs kind of scattered
throughout, and in various formats. Yet another new format this time
(AsciiDoc as implemented by AsciiDoctor).

Trying to give a short description of the major methods in ksb::Module
also made me realize that existing documentation on setModuleSet was
incorrect (claimed it used a module set name, but the actual code needed
a real ksb::ModuleSet). Added a check for that as well.