Try to streamline 'quick start' documentation.
This should document really almost everything the user needs to do to
get started with kdesrc-build.

The big thing missing is the .setup-env parts. I think that should go in
kdesrc-build-setup but that's a separate discussion.

Either way, once the documentation is accurate that should help in
determining which features to include in kdesrc-build to be able to make
this documentation even simpler so that we can broaden the base of users
able to be productive with this and related tools.