first-run: Add basic installer and packages for Debian distros.
This adds the installer to allow Debian-based distros to run and a basic
set of packages needed to at least allow kdesrc-build to successfully
run --initial-setup, as tested under a Docker image of debian 9.

This will still miss the dependencies required to fully make it through
any kind of reasonable KDE software build. In particular even if I
included the list of Qt modules that Debian 9 stable has available, they
are too old for current git-based applications. But other modules are
probably needed and still missing (suggestions accepted! :).

So there's still more to do here, which is why the bug remains open.

See issue #15

CCBUG:402901