Fix cxxflags being set to a space when globally empty.
When cxxflags is globally set to an empty value, the getOption magic
that appends module values to global values for cxxflags causes it the
result to equal ' ' (i.e. one space). This is because the space is a
separator between two empty values.

This causes code testing against cxxflags to think it's actually been
set to a value and to add it in the cmake calls.

I fix this and add a test case, but also add some insurance by trimming
leading/trailing white space so that the existing check for empty
cxxflags would have had a chance to catch this.

Differential Revision: https://phabricator.kde.org/D18165