git: Do not try to git-stash in git supermodules.
git-diff will show that there are differences in output in a git
supermodule (one containing active submodules). The fact that it does
this causes kdesrc-build to try to stash the local changes, although in
the vast majority of cases there will be no local *changes*, just local
differences in what commit is checked out for each submodule compared to
what the supermodule remote is expecting. This causes qt5 updates to
fail often in my testing.

I fix this by skipping git-stash for any git modules with submodules,
which allows the update to continue as normal.