env: Do not prepend /bin when {qt,kde}dir is blank.
This fixes #29, where kdesrc-build was aggressively prepending '/bin' to
the PATH if either kdedir or qtdir were unset (and qtdir is unset by
default). This isn't normally an issue but can conflict with user
settings.

Since the desire here is to setup Qt and/or KF5 before system packages,
only perform this prepending if Qt and/or KF5 are actually in use from
kdesrc-build. If those values aren't set then the system should already
be configured to provide in the environment so we shouldn't interfere.

It turns out that both ksb::Module and ksb::BuildContext are already
somewhat testable without a lot of orchestration needed here. I've
verified that the new tests fails without the patch applied, and pass
with this patch applied.