mojo: Correct testcase for bug 394497.
The intervening inclusion of the dependency-graph code changed how this
test must be implemented, but my initial merge failed to retain the
fixes to Mojo-based behavior that had already been included in the
make_it_mojo branch.

After this commit, all existing test cases now pass again (as of
Mojolicious 8.12), which will hopefully be reflected in CI once I update
the .gitlab-ci.yml to add Mojolicious.