mojo: Embed Mojolicious 8.19.
So after researching the best way to add a dependency to Mojolicious,
(without forcing users to also run their own perlbrew, local::lib, or
similar setup), I think the easiest/best solution to use of Mojolicious
within kdesrc-build is to embed it directly.

This is because the distro-provided Mojolicious I'm able to find (even
on distros that supply it) is sometimes too old, which causes a great
deal of kdesrc-build errors or even worse, causes kdesrc-build to
appeaer to try to work but silently fail instead.

Mojolicious is technically well-suited to this arrangement as it is pure
Perl (though it can use compiled Perl modules) and very svelte for the
features it provides, adding about 2.2 MB to the total repository size.

Duplicating the existing setup is relatively simple as well if you
already have a cpanminus setup, you run
    cpanm -n -L /tmp/perl Mojolicious

to have cpanminus install Mojolicious and non-core dependencies to
/tmp/perl, and copy out the /tmp/perl/lib/{Mojo/,Mojolicious/,}*.pm into
modules/ in the kdesrc-build source repository. You can also remove
things like the template JS, HTML templates, etc. Of course it is
probably worth scripting that too.

Mojolicious itself is distributed under the Artistic License.