build-system: Don't auto-clean build dir on repeated failure.
This was needed prior to CMake (and for some time as we were porting to
CMake) so that a wedged build system had at least a hope of eventually
being corrected.

This interferes with other use cases however and since CMake is really
good nowadays at generating valid build systems, remove this behavior.
Users can still pass --refresh-build when desired.

Fixes #30.