sample-rc: Give the only anon module-set a name.
It is optional for module-sets to have names, but it's still a good idea
to *have* a name since kdesrc-build will try to provide you information
on which module-set a given module is being built from. If no name is
present then you'll get an ugly name for the set like
"Building qca from <module-set at line 34>".

As long as the user doesn't mind that's fine, but our base configuration
should set a higher standard.

This was the only anonymous module-set in the current base
configuration. Thanks to Laurent for the report.

BUG:409198
FIXED-IN:19.08