git: Improve detection of local changes.
libgpg-error currently fails to update with kdesrc-build, ultimately due
to git-rebase complaining that there are local changes that need to be
stashed or reverted first.

These local changes are inherent to the libgpg-error build process, but
even so, kdesrc-build *does* check for local changes and stashes them
automatically.

But it turns out the current check from git of whether there are local
changes (using git-diff) gives a 'no changes' answer with libgpg-error's
modifications. So git-diff thinks there's no change, but git-rebase
thinks there is a change.

git-status --short gives an answer matching git-rebase, so switch to
that instead. git-status also gives answers for both the index and the
working tree at the same time to the updated code is even a bit shorter
now.