dep-resolv: Introduce --no-include-dependencies cmdline opt.
Does what it says on the tin, to permit disabling dependency inclusion
on a one-time basis without editing config file. This just involved
moving changing this option internally to be a 'global flag' which
already are automagically able to be negated by precedining with --no-.

When adding documentation for it I noticed I had not documented it in
the Docbook docs so I've added docs for that and --include-dependencies
also and updated copyright.

This should also support updating other documentation for the issue
being tracked at #31

Original commit: 6c10817e
https://invent.kde.org/sdk/kdesrc-build/-/commit/6c10817e55adcbb14ff17b5df8c224f5e416862c