mojo: Make 'kdesrc-build --backend' sort of work again.
This is the test client for the backend/client split. It's awful. But
you should at least be able to type in the modules to build (or select
from the dropdown), click the button to post the list to the backend,
and then get a "Start Build" style button once the backend returns with
the list of modules it thinks it will build in response.

TODO: Make --include-dependencies a checkbox, make update/build phases
configurable, make "live updates" that don't rely on re-reading full
rc-file each time, etc.