Introduce a DTO for exporting dependency information.
The structure of the raw dependency information spit out by ksb::DependencyResolver is not suitable to expose to clients directly.
It simply is too tightly coupled to Perl code and kdesrc-build in particular.

By introducing a DTO to define the wire format for exported information, we get much cleaner output that is unambiguous and can be exported safely to clients.