module_list: Fix ignoring module-sets.
kdesrc-build used to support doing things like
--ignore-modules $name_of_module_set to filter out entire module lists
easily. At some point it stopped working, possibly due to all the logic
involved in this (some of which is likely dead code and overdue for a
cleanup).

For now this fixes the issue in my testing, and I've added a test case
to help prevent recurrence in the future.

Fixes issue #36 on the Gitlab. There's no corresponding Bugzilla bug.