Fix ignore-modules breakage.
This resolves the failing test, with the cause being that
`expandModuleSets` (in ModuleResolver.pm) did *not* handle ignored
modules, only `resolveSelectorsIntoModules` did.

ksb::Application's `modulesFromSelectors` calls one or the other based
on whether the user had command line arguments to select a specific
subset of modules or not.

This is fixed by moving the check to `expandModuleSets` (which is called
from `resolveSelectorsIntoModules`), which also required permitting
a function to expand a single module set to expand to an empty list (if
all modules were ignored).