dep-resolv: Try to sort in rc-file order if not otherwise prevented.
This commit introduces a secondary build ordering to add to the ordering
criteria, so that modules would build approximately in the order seen
within the config file as defined by the user.

This permits the user some control in module ordering (as was previously
the case), while still building per the dependencies specified in
kde-build-metadata and prioritizing modules that are commonly depended
upon to build earlier.

To support situations where kdesrc-build is used without an rc-file we
still include the previous tiebreaker (sorting by module name) to ensure
a reproducible build order for a given (and unchanged) set of dependency
information and selectors.

Fixes #41.