git: Capture basic git-status info before error-prone ops.
There have long been bug reports around git-stash handling. I just had
to enter another one, but it's hard to troubleshoot what git was seeing
after the fetch and rebase has already happened, even if you try to
rewind with reflog.

For now this will just run a simple 'git-status' as a smoke check before
the update, and optionally if a stash or error occur. My hope is this
will permit more detail to be provided with subsequent bug reports.

CCBUG:420933