mojo: Try to make 'airport' display on module build more clear.
As suggested by Johan Ouwerkerk on !40, use more color on the result
line and also give a clear non-color indication. This should make the
output more useful at a glance at the TTY (though the goal here it to
allow a GUI or web frontend so I don't want to spend forever here
either).