git: Revert change to git-desired-protocol.
We *always* use HTTPS to fetch new objects from git.

We *can support* setting the git "push URL" to HTTPS also, and some
users will need this. But this complicates the existing practice of
using git + SSH to push changes to the KDE git infrastructure, so revert
this part of the change.