module: Add ability to track messages to show user after build ends.
This is intended to support some of the git-stash improvement we want in
issue #42 but it will obviously have wider application.

The hard part is that in the normal mode of operations (asynchronous)
the update process is off in a separate subprocess and will not have
access to the ksb::Module that we would want to print the messages from
at the end of the main process.

As with the similar problems I've run into here, my solution is to pass
the info over the IPC class. This will be easier in the make_it_mojo
branch but on the other hand it will need implemented differently in
that branch.