Ignore `hasrepo` metadata field.
This field is effectively obsolete, and sysadmin is looking to remove it.
Previously there used to be metadata files on the 'group' level that had
this flag disabled, but such files are now all gone. The sanity check
never triggers for the metadata files anymore since the move to Gitlab.