Add more os-release candidates.
On FreeBSD (prior to 13), os-release is installed by a port (so it
may be missing -- but it's an installation requirement for all KDE
and possibly Qt ports, so it is going to be there if whoever is using
kdesrc-build has any other KDE software at all). The file ends up
in */usr/local/etc* because */etc* is for base-system things.

Add that candidate location to the checks.