kde_projects: Adjust new location for dep/branch info (kde-build-metadata).
The sysadmins have migrated kde-build-metadata into repo-metadata (under
/dependencies for now). This helps with project management and reduces
the risk of having kdesrc-build in an inconsistent state based on when
repo-metadata is updated vs. kde-build-metadata.

Update kdesrc-build to look for the dependency graph and
branch/branch-group information in the updated location so that we can
adapt to the move.

This passes test suite and has been tested by an actual short set of
compile jobs here.