mojo: Simply subprocess option updates to parent, fix persistents.
Mojolicious is perfectly able to handle passing entire hashrefs across
the process boundary so just do the simple thing and copy the entire
options bundle for the module back to the parent.

This solves the issue of how to ensure ->setOption calls in the
subprocess are then reflected in the parent for the rest of the script
runtime.

This also makes it easy to fix the same things for persistent option
handling (which was recently fixed in master using the legacy IPC code),
so that is also done.