Build libdbusmenu-qt without documentation
Eliminates the only dependency on doxygen in the build, thus simplifying
the environment setup.

libdbusmenu-qt is also the only thing we pull from Launchpad with bzr.
Upstream appears to be dead, so it would be good to import the code into
invent.kde.org to further simply the build process.