qt5: Remove from default build (kdesrc-build-setup and --initial-setup).
With the advent of Qt6 development, the pace of Qt5 has slowed enough
that we can now rely on distribution packages being available and up to
date for Plasma development.

Since Qt5 remains a significant pain point for kdesrc-build users, we
shouldn't inflict this requirement by default for new users.