cmdline: Make _readCommandLineOptions return an object instead of ref param.
This is meant to allow me to refactor some of the processing that
happens while reading command line options to happen in a separate step.
The ultimate goal is to completely separate "reading the command line
options" from "setting the options used by ksb::Application" so that the
underlying Mojolicious application has no concept of command line
options at all.