mojo: Decouple 'early' options from web backend setup.
We now pass the t/smoke/install-and-run.t test case. Perhaps more
importantly, we can now in general ensure that cmdline option processing
which is anything other than idempotent can be handled exactly once,
rather than the two separate streams that occured between the backend
and the UI / main process.

Fixing up the test suite to adapt to yet another change to the way we
create the Application object is annoying, so I've added a helper
function for that. Likewise I noticed in testing that I wasn't seeing
the backtraces that would have been helpful under test so I've altered
croak_runtime to act as croak_internal when running under Test::Harness
(as with `prove`).