sample-config: Remove libdbusmenu-qt from default build.
For now it's only commented out but the module hasn't changed in many
years so there's really no need for kdesrc-build to have to build it. It
should be installed from -devel packages instead as we do with other
distro dependencies.

I've commented it out for now as an indicator this was deliberate, but
I've taken the liberty of updating the distros in ksb::FirstRun as
appears to be appropriate. After a few weeks we can remove the commented
section as dead code.

See issue #53.