xsession: Fix selection of wrong libname when a compat symlink is present.
The sample xsession/environment setup shell script has to figure how
where CMake is going to install the libraries that kdesrc-build builds,
so that it may setup the proper override paths in the environment.

These library names are a cornucopia of names across various Linux
distros. Some systems have /usr/lib, others use /usr/lib64, some use
both, some use /usr/lib and /usr/lib32, and still others use a combo of
all three, with /usr/lib having the "right" libraries and one of the
other specific paths being a symlink back to /usr/lib if appropriate.

In this latter situation our script guesses the wrong library name, with
predicable failures from there. This breaks Arch Linux if users don't
know to fix the session driver that kdesrc-build installs.

The fix here is simple, we only treat /usr/lib64 as the authoritative
library path if it isn't a symlink elsewhere. Otherwise we use /usr/lib
as normal and assume that CMake itself will also create a path ending in
/lib (and not /lib64 or /lib32).

Fixes #52, and my thanks to Vlad for the detailed bug report, steps to
reproduce, isolating the cause and testing the fix.