mojo: Fix persistent options across subprocesses again.
I noticed due to problems with saving the number of warnings from the
last build.

Ironically, *that* specific bug was a simple fix in BuildSystem.pm in
retrospect, but I noticed while trying to track it down that the
previous logic of simply overwriting our previous object stores holding
persistent options or runtime options could also cause problems, due to
the potential of wiping out a change that was made in the main process
(directly or by a different subprocess).

To improve this I change the subprocess handling code so that we apply
only the changes that occurred during the subprocess lifetime. This
doesn't completely eliminate the possibility of problems but since most
subprocesses operate on very distinct parts of the options namespaces
this should work for now.