mojo: Fix the --install option.
This finishes up for now on my quest to decouple some of the module
build phases apart.

Rather than completely scrapping the existing build handler, I realize
that it's already nearly in the intended form of acting as a
pre-build-steps, build phases, and post-build handler. So part of the
fix is simply to rename _handle_build to reflect its current role.

I did also have to adjust the Module's runPhase handler to skip phases
which aren't supposed to run. This is sufficient to make --install work
again.