mojo: Crash early if the rc-file is invalid.
Config files which are invalid (e.g. with a comment in front of a module
block where the '#' character was taken out) would recently have caused
crashes with no error output at all.

Recent changes I made for other reasons improved this somewhat, but in
the other direction, with a screenful of error spam, sometimes even
duplicated as the backend server would create or re-create the
ksb::Application multiple times before things finally settled out in an
HTTP 500 response to the U/I process.

Fix this by forcing the rc-file to be read before we go forward with
launching a backend server. Unfortunately it doesn't really make sense
to inject the backend server with a Perl object as I had been thinking
from the last commit, because we actually want the backend server to be
able to work potentially with multiple different build directories, each
with different rc-files, and driven by non-Perl clients. So the backend
will have to read the rc-file at some point.

But in the interactive mode where the backend server is just an
implementation detail, that can be made better, so I've done so.