mojo: Leave Mojo logging enabled as it is already going to a file.
I had squelched the Mojolicious logging output earlier when all such
output went to STDERR, but now that it's being directed to a file it's
better to have this output available (at a reasonable logging level, at
least). So stop squelching it and set it to default to the 'info' level.

At the same time add the ability to show debug output (both to console
and mojo-backend.log). As far as I can tell --debug Does the Right Thing
even under this Mojo-based solution now.