mojo: Remove needless 'setup' pre-build phase.
For starters I spelled this differently between the default phase names
in ksb::BuildContext and ksb::Module's default phase handlers.

But this is now unneeded anyways. It was intended to handle module build
phase setup tasks that were independent of whether a build/test/install
were performed or not, but since ksb::Module's "build" sub is sticking
around and is now always called, and it already does the right thing,
just let it do the right thing.