Require Perl 5.22 or better.
This was performed programmatically using the commands:
    find . '(' -name '*.t' -o -name '*.pm' -o -name '*.pl' ')' \
            -execdir sed -i -e 's/^use 5.014/use v5.22/' '{}' ';'
    find . '(' -name '*.t' -o -name '*.pm' -o -name '*.pl' ')' \
            -execdir sed -i -e 's/^use 5.018/use v5.22/' '{}' ';'

and then manual inspection from there after grepping for files to look
for any Perl code matching 'use [0-9]' to require a given Perl version.

The justification was provided in the updated README.md. TL;DR: Even
ancient distros support at least 5.22. If this is a problem for you,
please upgrade!