setup: Make include-dependencies optional, show module groups not chosen.
My hope is to make it easier for new users to understand how to add
module groups if they are desired later, since there has previously been
no generated note or warning in the generated kdesrc-buildrc if the user
chose no major groups.

In this case we had a user who, even though they didn't select any
module groups, still had kdesrc-build trying to build bluez-qt due to
include-dependencies being set to 'true' (the default). However bluez-qt
needs a CMake option to be passed to build with kdesrc-build; this
option isn't set unless the kf5-frameworks-build-include file is read
in.

That fix will happen in a separate commit, this should hopefully make it
easier to read through the generated kdesrc-buildrc in any case.

CCBUG:426219