mojo: Add test case for phase handling.
The idea will be to use this as a integrated test for the various ways
of turning off and on various build phases globally and for each module,
between cmdline options, global rc-file options and per-module rc-file
options.

Things are a way off and there are still TODOs to fix but it's better
than no test!