mojo: Add testcase for web server functionality.
This is very basic -- tests only whether the web server backend turns
on and reads the desired rc-file. On the other hand, working tests grow
better tests, and this is the first specialized to the backend/frontend
split I'm working to.

Since Mojolicious uses its own standard config mechanism I had to rewire
how the options get passed into the web::BackendServer but other than
that, this was mostly straightforward.